Switch target pace config to use a string
We now accept a string as the pace configuration, which can be a number or of
the form mm::ss. This required a custom parser, and I also added some
flexibility to allow comma decimal separators, which seem to not be handled
natively by the SDK. Unfortunately, this means that we will no longer get any
number validation in the app settings.

We also prefer mm:ss format for showing pace, or just the minutes if it's an
integer number of minutes.